requirements.txt
dxcam==0.0.5
PySide6==6.7.2
PySide6-Fluent-Widgets==1.6.0
pywin32==306
opencv-python==4.10.0.84
cnocr==2.3.0.3
PyAutoGUI==0.9.54